+++ examples/gateway-shard.rs
@@ -7,7 +7,7 @@ async fn main() -> anyhow::Result<()> {
     // Initialize the tracing subscriber.
     tracing_subscriber::fmt::init();
 
-    let intents = Intents::GUILD_MESSAGES | Intents::GUILD_VOICE_STATES;
+    let intents = Intents::GUILDS | Intents::GUILD_MODERATION;
     let (shard, mut events) = Shard::new(env::var("DISCORD_TOKEN")?, intents);
 
     shard.start().await?;