+++ twilight-cache-inmemory/src/event/message.rs
@@ -202,6 +202,7 @@ mod tests {
             pinned: false,
             reactions: Vec::new(),
             reference: None,
+            role_subscription_data: None,
             sticker_items: Vec::new(),
             thread: None,
             referenced_message: None,