+++ twilight-gateway/src/cluster/mod.rs
@@ -14,8 +14,9 @@
 //! #[tokio::main]
 //! async fn main() -> anyhow::Result<()> {
 //!     let token = env::var("DISCORD_TOKEN")?;
-//!     let intents =
-//!         Intents::GUILD_BANS | Intents::GUILD_EMOJIS_AND_STICKERS | Intents::GUILD_MESSAGES;
+//!     let intents = Intents::GUILD_MODERATION
+//!         | Intents::GUILD_EMOJIS_AND_STICKERS
+//!         | Intents::GUILD_MESSAGES;
 //!     let (cluster, mut events) = Cluster::new(token, intents).await?;
 //!     let cluster = Arc::new(cluster);
 //!     cluster.up().await;