+++ twilight-model/src/gateway/event/mod.rs
@@ -60,6 +60,8 @@ pub enum Event {
     GatewayReconnect,
     /// Undocumented event, should be ignored.
     GiftCodeUpdate,
+    /// A audit log entry was created.
+    GuildAuditLogEntryCreate(Box<GuildAuditLogEntryCreate>),
     /// A guild was created.
     GuildCreate(Box<GuildCreate>),
     /// A guild was deleted or the current user was removed from a guild.
@@ -205,6 +207,7 @@ impl Event {
             Event::ChannelDelete(e) => e.0.guild_id,
             Event::ChannelUpdate(e) => e.0.guild_id,
             Event::CommandPermissionsUpdate(e) => Some(e.0.guild_id),
+            Event::GuildAuditLogEntryCreate(e) => e.0.guild_id,
             Event::GuildCreate(e) => Some(e.0.id),
             Event::GuildDelete(e) => Some(e.id),
             Event::GuildEmojisUpdate(e) => Some(e.guild_id),
@@ -292,6 +295,7 @@ impl Event {
             Self::GatewayInvalidateSession(_) => EventType::GatewayInvalidateSession,
             Self::GatewayReconnect => EventType::GatewayReconnect,
             Self::GiftCodeUpdate => EventType::GiftCodeUpdate,
+            Self::GuildAuditLogEntryCreate(_) => EventType::GuildAuditLogEntryCreate,
             Self::GuildCreate(_) => EventType::GuildCreate,
             Self::GuildDelete(_) => EventType::GuildDelete,
             Self::GuildEmojisUpdate(_) => EventType::GuildEmojisUpdate,
@@ -371,6 +375,7 @@ impl From<DispatchEvent> for Event {
             DispatchEvent::ChannelUpdate(v) => Self::ChannelUpdate(v),
             DispatchEvent::CommandPermissionsUpdate(v) => Self::CommandPermissionsUpdate(v),
             DispatchEvent::GiftCodeUpdate => Self::GiftCodeUpdate,
+            DispatchEvent::GuildAuditLogEntryCreate(v) => Self::GuildAuditLogEntryCreate(v),
             DispatchEvent::GuildCreate(v) => Self::GuildCreate(v),
             DispatchEvent::GuildDelete(v) => Self::GuildDelete(v),
             DispatchEvent::GuildEmojisUpdate(v) => Self::GuildEmojisUpdate(v),