+++ twilight-model/src/guild/partial_guild.rs
@@ -45,6 +45,9 @@ pub struct PartialGuild {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub premium_subscription_count: Option<u64>,
     pub premium_tier: PremiumTier,
+    /// ID of the where moderators of Community guilds receive notices from
+    /// Discord.
+    pub public_updates_channel_id: Option<Id<ChannelMarker>>,
     pub roles: Vec<Role>,
     pub rules_channel_id: Option<Id<ChannelMarker>>,
     pub splash: Option<ImageHash>,
@@ -98,6 +101,7 @@ mod tests {
             premium_progress_bar_enabled: true,
             premium_subscription_count: Some(3),
             premium_tier: PremiumTier::Tier1,
+            public_updates_channel_id: None,
             roles: Vec::new(),
             rules_channel_id: Some(Id::new(6)),
             splash: Some(image_hash::SPLASH),
@@ -114,7 +118,7 @@ mod tests {
             &[
                 Token::Struct {
                     name: "PartialGuild",
-                    len: 34,
+                    len: 35,
                 },
                 Token::Str("id"),
                 Token::NewtypeStruct { name: "Id" },
@@ -185,6 +189,8 @@ mod tests {
                 Token::U64(3),
                 Token::Str("premium_tier"),
                 Token::U8(1),
+                Token::Str("public_updates_channel_id"),
+                Token::None,
                 Token::Str("roles"),
                 Token::Seq { len: Some(0) },
                 Token::SeqEnd,