+++ twilight-util/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.14.2] - 2023-01-20
+
+### Features
+
+- add role subscriptions ([#2034](https://github.com/twilight-rs/twilight/issues/2034))
+
 ## [0.14.1] - 2023-01-07
 
 ### Features